Removed  navbar from login and other authentication pages
