+++ src/screens/authenticated/dashboard/index.tsx
@@ -3,6 +3,8 @@ import { Button } from 'react-bootstrap'
 import { getAuth, signOut } from 'firebase/auth'
 import { useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
+import SideBar from "../dashboard/sub-components/navbar/index";
+
 
 const Dashboard = () => {
   const navigate = useNavigate()
@@ -14,19 +16,10 @@ const Dashboard = () => {
     }
   }, []);
   
-  const handleSignOut = () =>{
-    const auth = getAuth()
-    signOut(auth)
-    sessionStorage.clear()
-    console.log('Deleted token')
-    console.log('Redirecting to login page')
-    navigate('/')
-  }
-
   return (
+    
     <div>
       <h1>DASHBOARD</h1>
-    <Button onClick={handleSignOut}>Logout</Button>
     </div>
   )
 }
